Extract compact layout, medium layout, and shared components into their own files.
